backend/requirements.txt
# ClaraCare Backend Dependencies

# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# WebSockets
websockets==13.1
python-dotenv==1.0.1

# HTTP Client
httpx==0.27.2

# Optional (will add when needed)
# sanity-client==1.0.0  # For Sanity CMS integration
# pydantic==2.9.0  # Data validation
